Refactor comment option utilities into shared options module
